+++ src/app/model/Search/Filter/CriteriaSearchFilter.ts
