+++ src/app/model/Search/Filter/CriteriaSearchFilterValue.ts
